# server: Cache PreparedEvalQuery's for data requests