@patrick-east 2019-12-13
Some performance details.. I tested it out with big data sets and pretty substantial amounts of modules, the memory increase is pretty negligible with the cached queries. We could probably make the cap quite a bit higher without any issues but I seriously doubt anyone needs more than 100.

I saw on the e2e authz REST API benchmark that it shaves off a pretty decent chunk of time per request, something like 200us (~30-40% of overall time on the smaller inputs requests), and as expected cuts down on memory usage per request:

Before:
```
BenchmarkRESTAuthzForbidAuthn-8      	    3171	    378555 ns/op	  111811 B/op	    2047 allocs/op
BenchmarkRESTAuthzForbidPath-8       	    2821	    418248 ns/op	  127505 B/op	    2465 allocs/op
BenchmarkRESTAuthzForbidMethod-8     	    2990	    418550 ns/op	  128378 B/op	    2477 allocs/op
BenchmarkRESTAuthzAllow10Paths-8     	    2862	    414088 ns/op	  127129 B/op	    2461 allocs/op
BenchmarkRESTAuthzAllow100Paths-8    	    1468	    715536 ns/op	  248767 B/op	    5979 allocs/op
BenchmarkRESTAuthzAllow1000Paths-8   	     256	   3956631 ns/op	 1486013 B/op	   41988 allocs/op
```

After: 
```
BenchmarkRESTAuthzForbidAuthn-8      	    6580	    156077 ns/op	   34929 B/op	     463 allocs/op
BenchmarkRESTAuthzForbidPath-8       	    5658	    198001 ns/op	   50306 B/op	     882 allocs/op
BenchmarkRESTAuthzForbidMethod-8     	    5401	    212781 ns/op	   51045 B/op	     894 allocs/op
BenchmarkRESTAuthzAllow10Paths-8     	    5814	    201236 ns/op	   49882 B/op	     878 allocs/op
BenchmarkRESTAuthzAllow100Paths-8    	    2559	    497152 ns/op	  173028 B/op	    4397 allocs/op
BenchmarkRESTAuthzAllow1000Paths-8   	     306	   3629313 ns/op	 1410577 B/op	   40410 allocs/op
```